@@ -3,11 +3,11 @@ package com.giova.service.moneystats.app.stats;
 import com.giova.service.moneystats.app.stats.dto.Stats;
 import com.giova.service.moneystats.app.stats.entity.StatsEntity;
 import com.giova.service.moneystats.app.wallet.entity.WalletEntity;
-import com.giova.service.moneystats.authentication.entity.UserEntity;
+import com.giova.service.moneystats.authentication.dto.UserData;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
+import io.github.giovannilamarmora.utils.utilities.ObjectToolkit;
 import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Component;
 
@@ -15,22 +15,8 @@ import org.springframework.stereotype.Component;
 public class StatsMapper {
 
   @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
-  public List<StatsEntity> fromStatsToEntity(
-      List<Stats> stats, WalletEntity wallet, UserEntity user) {
-    return stats.stream()
-        .map(
-            stats1 -> {
-              StatsEntity statsEntity = new StatsEntity();
-              BeanUtils.copyProperties(stats1, statsEntity);
-              statsEntity.setWallet(wallet);
-              statsEntity.setUser(user);
-              return statsEntity;
-            })
-        .collect(Collectors.toList());
-  }
-
-  @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
-  public List<Stats> fromEntityToStats(List<StatsEntity> statsEntities) {
+  public static List<Stats> fromEntityToStats(List<StatsEntity> statsEntities) {
+    if (ObjectToolkit.isNullOrEmpty(statsEntities)) return null;
     return statsEntities.stream()
         .map(
             statsEntity -> {
@@ -38,6 +24,22 @@ public class StatsMapper {
               BeanUtils.copyProperties(statsEntity, stats2);
               return stats2;
             })
-        .collect(Collectors.toList());
+        .toList();
+  }
+
+  @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
+  public static List<StatsEntity> fromStatsToEntity(
+      List<Stats> stats, WalletEntity wallet, UserData user) {
+    if (ObjectToolkit.isNullOrEmpty(stats)) return null;
+    return stats.stream()
+        .map(
+            stats1 -> {
+              StatsEntity statsEntity = new StatsEntity();
+              BeanUtils.copyProperties(stats1, statsEntity);
+              statsEntity.setWallet(wallet);
+              statsEntity.setUserIdentifier(user.getIdentifier());
+              return statsEntity;
+            })
+        .toList();
   }
 }
