@@ -1,23 +1,20 @@
 package com.giova.service.moneystats.crypto;
 
-import com.giova.service.moneystats.scheduler.CronCachingReset;
-import com.giova.service.moneystats.scheduler.CronMarketData;
 import io.github.giovannilamarmora.utils.exception.dto.ExceptionResponse;
 import io.github.giovannilamarmora.utils.generic.Response;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
 import io.github.giovannilamarmora.utils.interceptors.Logged;
-import io.github.giovannilamarmora.utils.interceptors.correlationID.CorrelationIdUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import javax.validation.Valid;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -30,8 +27,6 @@ import org.springframework.web.bind.annotation.*;
 public class CryptoController {
 
   @Autowired private CryptoService appService;
-  @Autowired private CronMarketData cronMarketData;
-  @Autowired private CronCachingReset cronCachingReset;
 
   @GetMapping(value = "/dashboard", produces = MediaType.APPLICATION_JSON_VALUE)
   @Operation(
@@ -48,28 +43,22 @@ public class CryptoController {
               examples = @ExampleObject(value = "@crypto-dashboard.json")))
   @ApiResponse(
       responseCode = "401",
-      description = "Invalid JWE",
+      description = "Invalid Token",
       content =
           @Content(
               schema = @Schema(implementation = ExceptionResponse.class),
               mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@invalid-jwe.json")))
-  @ApiResponse(
-      responseCode = "401",
-      description = "Expired JWE",
-      content =
-          @Content(
-              schema = @Schema(implementation = ExceptionResponse.class),
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@expired-jwe.json")))
+              examples = @ExampleObject(value = "@invalid-token-exception.json")))
   @LogInterceptor(type = LogTimeTracker.ActionType.CONTROLLER)
   public ResponseEntity<Response> getCryptoDashboard(
-      @RequestHeader(HttpHeaders.AUTHORIZATION) @Valid @Schema(description = "Authorization Token")
-          String authToken) {
+      @RequestHeader(HttpHeaders.AUTHORIZATION)
+          @Valid
+          @Schema(description = "Authorization Token", example = "Bearer eykihugUiOj6bihiguu...")
+          String token) {
     return appService.getCryptoDashboardData();
   }
 
-  @GetMapping(value = "/resume", produces = MediaType.APPLICATION_JSON_VALUE)
+  @GetMapping(value = "/resume/{year}", produces = MediaType.APPLICATION_JSON_VALUE)
   @Operation(description = "API to get Crypto Resume", summary = "Crypto Resume", tags = "Crypto")
   @ApiResponse(
       responseCode = "200",
@@ -81,29 +70,28 @@ public class CryptoController {
               examples = @ExampleObject(value = "@crypto-resume.json")))
   @ApiResponse(
       responseCode = "401",
-      description = "Invalid JWE",
+      description = "Invalid Token",
       content =
           @Content(
               schema = @Schema(implementation = ExceptionResponse.class),
               mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@invalid-jwe.json")))
-  @ApiResponse(
-      responseCode = "401",
-      description = "Expired JWE",
-      content =
-          @Content(
-              schema = @Schema(implementation = ExceptionResponse.class),
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@expired-jwe.json")))
+              examples = @ExampleObject(value = "@invalid-token-exception.json")))
   @LogInterceptor(type = LogTimeTracker.ActionType.CONTROLLER)
   public ResponseEntity<Response> getCryptoResume(
-      @RequestHeader(HttpHeaders.AUTHORIZATION) @Valid @Schema(description = "Authorization Token")
-          String authToken) {
-    return appService.getCryptoResumeData();
+      @RequestHeader(HttpHeaders.AUTHORIZATION)
+          @Valid
+          @Schema(description = "Authorization Token", example = "Bearer eykihugUiOj6bihiguu...")
+          String token,
+      @PathVariable(value = "year")
+          @Schema(description = "Year of the data to be searched", example = "2024")
+          @Valid
+          @NotNull(message = "Year is a required field")
+          Long year) {
+    return appService.getCryptoResumeData(year);
   }
 
-  @PatchMapping(value = "/marketData/import", produces = MediaType.APPLICATION_JSON_VALUE)
-  @Operation(description = "API to import MarketData", tags = "Crypto")
+  @GetMapping(value = "/history", produces = MediaType.APPLICATION_JSON_VALUE)
+  @Operation(description = "API to get Crypto Resume", summary = "Crypto Resume", tags = "Crypto")
   @ApiResponse(
       responseCode = "200",
       description = "Successful operation",
@@ -111,74 +99,21 @@ public class CryptoController {
           @Content(
               schema = @Schema(implementation = Response.class),
               mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@import-marketData.json")))
+              examples = @ExampleObject(value = "@crypto-resume.json")))
   @ApiResponse(
       responseCode = "401",
-      description = "Invalid JWE",
+      description = "Invalid Token",
       content =
           @Content(
               schema = @Schema(implementation = ExceptionResponse.class),
               mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@invalid-jwe.json")))
-  @ApiResponse(
-      responseCode = "401",
-      description = "Expired JWE",
-      content =
-          @Content(
-              schema = @Schema(implementation = ExceptionResponse.class),
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@expired-jwe.json")))
+              examples = @ExampleObject(value = "@invalid-token-exception.json")))
   @LogInterceptor(type = LogTimeTracker.ActionType.CONTROLLER)
-  public ResponseEntity<Response> importMarketData(
-      @RequestHeader(HttpHeaders.AUTHORIZATION) @Valid @Schema(description = "Authorization Token")
-          String authToken) {
-    cronMarketData.scheduleAllCryptoAsset();
-    Response response =
-        new Response(
-            HttpStatus.OK.value(),
-            "MarketData Successfully imported",
-            CorrelationIdUtils.getCorrelationId(),
-            null);
-    return ResponseEntity.ok(response);
-  }
-
-  @PatchMapping(value = "/cache/clean", produces = MediaType.APPLICATION_JSON_VALUE)
-  @Operation(description = "API to get Crypto Resume", summary = "Cache Clean", tags = "Crypto")
-  @ApiResponse(
-      responseCode = "200",
-      description = "Successful operation",
-      content =
-          @Content(
-              schema = @Schema(implementation = Response.class),
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@cache-clean.json")))
-  @ApiResponse(
-      responseCode = "401",
-      description = "Invalid JWE",
-      content =
-          @Content(
-              schema = @Schema(implementation = ExceptionResponse.class),
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@invalid-jwe.json")))
-  @ApiResponse(
-      responseCode = "401",
-      description = "Expired JWE",
-      content =
-          @Content(
-              schema = @Schema(implementation = ExceptionResponse.class),
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              examples = @ExampleObject(value = "@expired-jwe.json")))
-  @LogInterceptor(type = LogTimeTracker.ActionType.CONTROLLER)
-  public ResponseEntity<Response> cleanCache(
-      @RequestHeader(HttpHeaders.AUTHORIZATION) @Valid @Schema(description = "Authorization Token")
-          String authToken) {
-    cronCachingReset.scheduleCleanCache();
-    Response response =
-        new Response(
-            HttpStatus.OK.value(),
-            "Cache cleaned successfully",
-            CorrelationIdUtils.getCorrelationId(),
-            null);
-    return ResponseEntity.ok(response);
+  public ResponseEntity<Response> getCryptoHistory(
+      @RequestHeader(HttpHeaders.AUTHORIZATION)
+          @Valid
+          @Schema(description = "Authorization Token", example = "Bearer eykihugUiOj6bihiguu...")
+          String token) {
+    return appService.getCryptoHistoryData();
   }
 }
