@@ -1,31 +1,31 @@
 package com.giova.service.moneystats.crypto;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.giova.service.moneystats.app.stats.StatsService;
+import com.giova.service.moneystats.app.stats.StatsComponent;
 import com.giova.service.moneystats.app.stats.dto.Stats;
 import com.giova.service.moneystats.app.wallet.WalletService;
 import com.giova.service.moneystats.app.wallet.dto.Wallet;
-import com.giova.service.moneystats.authentication.entity.UserEntity;
+import com.giova.service.moneystats.authentication.dto.UserData;
+import com.giova.service.moneystats.crypto.asset.AssetMapper;
 import com.giova.service.moneystats.crypto.asset.AssetService;
 import com.giova.service.moneystats.crypto.asset.dto.Asset;
-import com.giova.service.moneystats.crypto.coinGecko.MarketDataService;
-import com.giova.service.moneystats.crypto.coinGecko.dto.MarketData;
+import com.giova.service.moneystats.crypto.marketData.MarketDataService;
+import com.giova.service.moneystats.crypto.marketData.dto.MarketData;
 import com.giova.service.moneystats.crypto.model.CryptoDashboard;
 import com.giova.service.moneystats.crypto.operations.dto.Operations;
-import io.github.giovannilamarmora.utils.exception.UtilsException;
+import io.github.giovannilamarmora.utils.context.TraceUtils;
 import io.github.giovannilamarmora.utils.generic.Response;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
 import io.github.giovannilamarmora.utils.interceptors.Logged;
-import io.github.giovannilamarmora.utils.interceptors.correlationID.CorrelationIdUtils;
 import io.github.giovannilamarmora.utils.math.MathService;
+import io.github.giovannilamarmora.utils.utilities.Mapper;
+import io.github.giovannilamarmora.utils.utilities.ObjectToolkit;
 import java.time.LocalDate;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,76 +42,90 @@ import org.springframework.util.ObjectUtils;
 public class CryptoService {
 
   private static final String BTC_SYMBOL = "BTC";
-  private final UserEntity user;
+  private final UserData user;
   private final Logger LOG = LoggerFactory.getLogger(this.getClass());
-  private final ObjectMapper objectMapper = new ObjectMapper().findAndRegisterModules();
   @Autowired private WalletService walletService;
-  @Autowired private StatsService statsService;
-  @Autowired private CryptoMapper cryptoMapper;
-  @Autowired private MarketDataService marketDataService;
   @Autowired private AssetService assetService;
+  @Autowired private StatsComponent statsComponent;
+  @Autowired private MarketDataService marketDataService;
 
   @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
   public ResponseEntity<Response> getCryptoDashboardData() {
     List<MarketData> marketData =
         marketDataService.getMarketData(user.getSettings().getCryptoCurrency());
 
-    List<LocalDate> getAllDates = statsService.getCryptoDistinctDates(user);
-    List<LocalDate> filter = new ArrayList<>();
-    Map<String, CryptoDashboard> getData = new HashMap<>();
-    int thisYear = 0;
+    List<LocalDate> getAllDates = statsComponent.getCryptoDistinctDates(user);
+    List<LocalDate> filter;
+    CryptoDashboard dashboard = new CryptoDashboard();
     if (!getAllDates.isEmpty()) {
-      int currentYear = getAllDates.get(getAllDates.size() - 1).getYear();
-      filter =
-          getAllDates.stream().filter(d -> d.getYear() == currentYear).collect(Collectors.toList());
+      int currentYear = getAllDates.getLast().getYear();
+      filter = getAllDates.stream().filter(d -> d.getYear() == currentYear).toList();
 
-      thisYear = currentYear;
-      getData = mapDashBoard(filter, marketData, false);
+      Map<String, CryptoDashboard> getData = mapDashBoard(filter, marketData, false);
+      if (!ObjectToolkit.isNullOrEmpty(getData)) {
+        dashboard = getData.get(String.valueOf(currentYear));
+      }
     } else {
-      List<Wallet> getAllWallet =
-          objectMapper.convertValue(
-              walletService.getCryptoWallets(true).getBody().getData(),
-              new TypeReference<List<Wallet>>() {});
-      CryptoDashboard dashboard =
-          cryptoMapper.mapCryptoDashboardWithoutStats(
+      List<Wallet> getAllWallet = new ArrayList<>();
+      ResponseEntity<Response> responseEntityWallet = walletService.getCryptoWallets(false);
+      if (!ObjectToolkit.isNullOrEmpty(responseEntityWallet.getBody())
+          & !ObjectToolkit.isNullOrEmpty(responseEntityWallet.getBody().getData()))
+        getAllWallet =
+            Mapper.convertObject(
+                responseEntityWallet.getBody().getData(), new TypeReference<List<Wallet>>() {});
+      dashboard =
+          CryptoMapper.mapCryptoDashboardWithoutStats(
               user.getSettings().getCryptoCurrency(),
               getAllWallet,
-              getCryptoAsset(false),
+              getCryptoAssetsList(false, null, getAllWallet, marketData, null),
               getAssetValue(marketData, BTC_SYMBOL));
-      getData.put(String.valueOf(thisYear), dashboard);
     }
+
+    dashboard.setWallets(null);
 
     String message = "Data for Crypto Dashboard!";
 
     Response response =
-        new Response(
-            HttpStatus.OK.value(),
-            message,
-            CorrelationIdUtils.getCorrelationId(),
-            getData.get(String.valueOf(thisYear)));
+        new Response(HttpStatus.OK.value(), message, TraceUtils.getSpanID(), dashboard);
     return ResponseEntity.ok(response);
   }
 
   @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
-  public ResponseEntity<Response> getCryptoResumeData() {
+  public ResponseEntity<Response> getCryptoResumeData(Long year) {
     List<MarketData> marketData =
         marketDataService.getMarketData(user.getSettings().getCryptoCurrency());
-    List<LocalDate> getAllDates = statsService.getCryptoDistinctDates(user);
+    List<LocalDate> getAllDates = statsComponent.getCryptoDistinctDates(user);
     Map<String, CryptoDashboard> getData = new HashMap<>();
+    CryptoDashboard dashboard = new CryptoDashboard();
     int thisYear = LocalDate.now().getYear();
     if (!getAllDates.isEmpty()) {
-      getData = mapDashBoard(getAllDates, marketData, true);
+      List<LocalDate> filterDateByYear =
+          getAllDates.stream().filter(localDate -> localDate.getYear() == year).toList();
+      Map<String, CryptoDashboard> getDataProvision =
+          mapDashBoard(filterDateByYear, marketData, true);
+      if (!ObjectToolkit.isNullOrEmpty(getDataProvision)) {
+        dashboard = getDataProvision.get(String.valueOf(year));
+      }
+      dashboard.setYearsWalletStats(
+          getAllDates.stream()
+              .map(LocalDate::getYear)
+              .distinct()
+              .sorted(Collections.reverseOrder())
+              .toList());
+      getData.put(String.valueOf(year), dashboard);
     } else {
-
-      List<Wallet> getAllWallet =
-          objectMapper.convertValue(
-              walletService.getCryptoWallets(false).getBody().getData(),
-              new TypeReference<List<Wallet>>() {});
-      CryptoDashboard dashboard =
-          cryptoMapper.mapCryptoDashboardWithoutStats(
+      List<Wallet> getAllWallet = new ArrayList<>();
+      ResponseEntity<Response> responseEntityWallet = walletService.getCryptoWallets(false);
+      if (!ObjectToolkit.isNullOrEmpty(responseEntityWallet.getBody())
+          & !ObjectToolkit.isNullOrEmpty(responseEntityWallet.getBody().getData()))
+        getAllWallet =
+            Mapper.convertObject(
+                responseEntityWallet.getBody().getData(), new TypeReference<List<Wallet>>() {});
+      dashboard =
+          CryptoMapper.mapCryptoDashboardWithoutStats(
               user.getSettings().getCryptoCurrency(),
               getAllWallet,
-              getCryptoAsset(true),
+              getCryptoAssetsList(false, null, getAllWallet, marketData, null),
               getAssetValue(marketData, BTC_SYMBOL));
       getData.put(String.valueOf(thisYear), dashboard);
     }
@@ -119,38 +133,107 @@ public class CryptoService {
     String message = "Data for Crypto Resume!";
 
     Response response =
-        new Response(
-            HttpStatus.OK.value(), message, CorrelationIdUtils.getCorrelationId(), getData);
+        new Response(HttpStatus.OK.value(), message, TraceUtils.getSpanID(), getData);
+    return ResponseEntity.ok(response);
+  }
+
+  @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
+  public ResponseEntity<Response> getCryptoHistoryData() {
+    LOG.info("Starting getting history data crypto for user {}", user.getUsername());
+
+    Map<String, CryptoDashboard> dashboardHashMap = new LinkedHashMap<>();
+    List<LocalDate> getAllDates = statsComponent.getCryptoDistinctDates(user);
+    List<Integer> distinctYears =
+        new ArrayList<>(
+            getAllDates.stream()
+                .map(LocalDate::getYear)
+                .distinct()
+                .sorted(Collections.reverseOrder())
+                .toList());
+
+    LocalDate today = LocalDate.now();
+    if (!distinctYears.contains(today.getYear())) distinctYears.add(today.getYear());
+    distinctYears.sort(Collections.reverseOrder());
+
+    ResponseEntity<Response> responseAssets = assetService.getAssets(false);
+    List<Asset> assets =
+        Optional.ofNullable(responseAssets.getBody())
+            .map(Response::getData)
+            .map(data -> Mapper.convertObject(data, new TypeReference<List<Asset>>() {}))
+            .orElse(Collections.emptyList());
+
+    AtomicInteger yearIndex = new AtomicInteger(0);
+
+    distinctYears.forEach(
+        year -> {
+          List<LocalDate> datesByYear =
+              getAllDates.stream().filter(d -> d.getYear() == year).toList();
+          CryptoDashboard dashboard = new CryptoDashboard();
+          AtomicReference<Double> balance = new AtomicReference<>(0D);
+
+          assets.forEach(
+              asset -> {
+                List<Stats> yearlyStats =
+                    Optional.ofNullable(asset.getHistory())
+                        .map(
+                            history ->
+                                history.stream()
+                                    .filter(h -> h.getDate().getYear() == year)
+                                    .toList())
+                        .orElse(Collections.emptyList());
+
+                if (yearIndex.get() == 0) {
+                  balance.updateAndGet(b -> b + asset.getValue());
+                }
+
+                if (!yearlyStats.isEmpty()) {
+                  if (yearIndex.get() != 0) {
+                    CryptoMapper.updateBalance(
+                        List.of(yearlyStats.getLast()), datesByYear, balance);
+                  }
+                }
+              });
+
+          dashboard.setBalance(MathService.round(balance.get(), 2));
+          dashboardHashMap.put(String.valueOf(year), dashboard);
+          yearIndex.incrementAndGet();
+        });
+
+    String message = "Data for History!";
+    Response response =
+        new Response(HttpStatus.OK.value(), message, TraceUtils.getSpanID(), dashboardHashMap);
     return ResponseEntity.ok(response);
   }
 
   private Map<String, CryptoDashboard> mapDashBoard(
-      List<LocalDate> dates, List<MarketData> marketData, Boolean isResume) throws UtilsException {
+      List<LocalDate> dates, List<MarketData> marketData, Boolean isResume) {
     Map<String, CryptoDashboard> response = new HashMap<>();
 
-    List<Integer> distinctDatesByYear =
-        dates.stream().map(LocalDate::getYear).distinct().collect(Collectors.toList());
+    List<Integer> distinctDatesByYear = dates.stream().map(LocalDate::getYear).distinct().toList();
 
-    // Wallet List
-    List<Wallet> getAllWallet =
-        objectMapper.convertValue(
-            walletService.getCryptoWallets(true).getBody().getData(),
-            new TypeReference<List<Wallet>>() {});
+    List<Wallet> getAllWallet = new ArrayList<>();
+    ResponseEntity<Response> responseEntityWallet = walletService.getCryptoWallets(true);
+    if (!ObjectToolkit.isNullOrEmpty(responseEntityWallet.getBody())
+        & !ObjectToolkit.isNullOrEmpty(responseEntityWallet.getBody().getData()))
+      getAllWallet =
+          Mapper.convertObject(
+              responseEntityWallet.getBody().getData(), new TypeReference<List<Wallet>>() {});
 
     AtomicInteger index = new AtomicInteger(0);
+    List<Wallet> finalGetAllWallet = getAllWallet;
     distinctDatesByYear.stream()
         .sorted(Collections.reverseOrder())
-        .peek(
+        .forEach(
             year -> {
               LOG.info("Mapping Data for year {}", year);
               // Filtro le date secondo l'anno
               List<LocalDate> filterDateByYear =
-                  dates.stream().filter(d -> d.getYear() == year).collect(Collectors.toList());
+                  dates.stream().filter(d -> d.getYear() == year).toList();
               CryptoDashboard dashboard = new CryptoDashboard();
-              dashboard.setLastUpdate(filterDateByYear.get(filterDateByYear.size() - 1));
+              dashboard.setLastUpdate(filterDateByYear.getLast());
               dashboard.setStatsAssetsDays(filterDateByYear);
               dashboard.setCurrency(user.getSettings().getCryptoCurrency());
-              dashboard.setPerformanceSince(filterDateByYear.get(0));
+              dashboard.setPerformanceSince(filterDateByYear.getFirst());
 
               AtomicReference<Double> balance = new AtomicReference<>(0D);
               AtomicReference<Double> initialBalance = new AtomicReference<>(0D);
@@ -163,25 +246,28 @@ public class CryptoService {
               AtomicInteger indexWallet = new AtomicInteger(0);
 
               List<Wallet> filterWallet =
-                  filterHistoryWallet(
-                      getAllWallet,
-                      balance,
-                      holdingBalance,
-                      initialBalance,
-                      lastBalance,
-                      holdingLastBalance,
-                      index,
-                      indexWallet,
-                      filterDateByYear,
-                      year,
-                      tradingBalance,
-                      tradingLastBalance);
+                  new ArrayList<>(
+                      filterHistoryWallet(
+                          finalGetAllWallet,
+                          balance,
+                          holdingBalance,
+                          initialBalance,
+                          lastBalance,
+                          holdingLastBalance,
+                          index,
+                          indexWallet,
+                          filterDateByYear,
+                          year,
+                          tradingBalance,
+                          tradingLastBalance));
 
               // dashboard.setAssets(getCryptoAsset(filterWallet, marketData, isResume));
-              dashboard.setAssets(getCryptoAssetByYear(isResume, year));
+              dashboard.setAssets(
+                  getCryptoAssetsList(isResume, year, filterWallet, marketData, dates));
 
               // Filtro Wallet cancellati da anni che non hanno stats
-              Predicate<Wallet> walletRemovedInThePast = wallet -> wallet.getDeletedDate() != null;
+              Predicate<Wallet> walletRemovedInThePast =
+                  wallet -> !ObjectToolkit.isNullOrEmpty(wallet.getDeletedDate());
               filterWallet.removeIf(walletRemovedInThePast);
 
               // Mi serve per mappare il passato
@@ -191,32 +277,25 @@ public class CryptoService {
                     wallet ->
                         wallet.getAssets() != null
                             && wallet.getAssets().stream()
-                                .filter(
-                                    asset ->
-                                        asset.getHistory() == null || asset.getHistory().isEmpty())
-                                .collect(Collectors.toList())
+                                .filter(asset -> !ObjectToolkit.isNullOrEmpty(asset.getHistory()))
+                                .toList()
                                 .isEmpty();
                 filterWallet.removeIf(hasEmptyStats);
               }
 
-              try {
-                dashboard.setWallets(filterWallet);
-                cryptoMapper.mapDashboardBalanceAndPerformance(
-                    dashboard,
-                    balance,
-                    holdingBalance,
-                    holdingLastBalance,
-                    tradingBalance,
-                    tradingLastBalance,
-                    getAssetValue(marketData, BTC_SYMBOL),
-                    marketData);
-              } catch (UtilsException e) {
-                throw new RuntimeException(e);
-              }
+              dashboard.setWallets(filterWallet);
+              CryptoMapper.mapDashboardBalanceAndPerformance(
+                  dashboard,
+                  balance,
+                  holdingBalance,
+                  holdingLastBalance,
+                  tradingBalance,
+                  tradingLastBalance,
+                  getAssetValue(marketData, BTC_SYMBOL),
+                  marketData);
               index.incrementAndGet();
               response.put(String.valueOf(year), dashboard);
-            })
-        .collect(Collectors.toList());
+            });
 
     return response;
   }
@@ -241,11 +320,11 @@ public class CryptoService {
               Wallet wallet1 = new Wallet();
               BeanUtils.copyProperties(wallet, wallet1);
               if (wallet.getHistory() != null) {
-                Stats history = wallet.getHistory().get(wallet.getHistory().size() - 1);
+                Stats history = wallet.getHistory().getLast();
                 wallet1.setHistory(List.of(history));
               }
 
-              if (wallet.getAssets() != null)
+              if (!ObjectToolkit.isNullOrEmpty(wallet.getAssets()))
                 wallet1.setAssets(
                     wallet.getAssets().stream()
                         .map(
@@ -253,11 +332,11 @@ public class CryptoService {
                               Asset asset1 = new Asset();
                               BeanUtils.copyProperties(asset, asset1);
                               List<Stats> listFilter = new ArrayList<>();
-                              if (asset.getHistory() != null) {
+                              if (!ObjectToolkit.isNullOrEmpty(asset.getHistory())) {
                                 listFilter =
                                     asset.getHistory().stream()
                                         .filter(h -> h.getDate().getYear() == year)
-                                        .collect(Collectors.toList());
+                                        .toList();
 
                                 asset1.setHistory(listFilter);
                               }
@@ -276,16 +355,16 @@ public class CryptoService {
                                 balance.updateAndGet(v -> v + asset1.getValue());
                               if (!listFilter.isEmpty()) {
                                 if (index.get() != 0)
-                                  cryptoMapper.updateBalance(listFilter, filterDateByYear, balance);
+                                  CryptoMapper.updateBalance(listFilter, filterDateByYear, balance);
                                 // if (index.get() == 0)
                                 //  balance.updateAndGet(v -> v + asset1.getValue());
                                 // else
                                 //  cryptoMapper.updateBalance(listFilter, filterDateByYear,
                                 // balance);
-                                cryptoMapper.updateInitialBalance(
+                                CryptoMapper.updateInitialBalance(
                                     listFilter, filterDateByYear, initialBalance);
 
-                                cryptoMapper.updateLastBalance(
+                                CryptoMapper.updateLastBalance(
                                     listFilter,
                                     filterDateByYear,
                                     lastBalance,
@@ -295,9 +374,43 @@ public class CryptoService {
 
                               checkAndMapWalletInThePast(
                                   index, listFilter, filterDateByYear, wallet1);
+
+                              if (!ObjectToolkit.isNullOrEmpty(asset1.getOperations())) {
+                                asset1.setOperations(
+                                    asset1.getOperations().stream()
+                                        .filter(
+                                            o -> {
+                                              boolean isExitDateNull =
+                                                  ObjectToolkit.isNullOrEmpty(o.getExitDate());
+                                              if (isExitDateNull) {
+                                                return o.getEntryDate().getYear() <= year;
+                                              } else {
+                                                return o.getExitDate().getYear() == year;
+                                              }
+                                            })
+                                        .toList());
+                              }
                               return asset1;
+
+                              //  if (!ObjectToolkit.isNullOrEmpty(asset1.getOperations()))
+                              //  asset1.setOperations(
+                              //      new ArrayList<>(asset1.getOperations())
+                              //          .stream()
+                              //              .filter(
+                              //                  o ->
+                              //                      //
+                              // (ObjectToolkit.isNullOrEmpty(o.getExitDate())
+                              //                      //        && o.getEntryDate().getYear() <=
+                              // year)
+                              //                      //    || o.getEntryDate().getYear() == year)
+                              //                      // Inverted Entry to Exit date
+                              //                      (ObjectToolkit.isNullOrEmpty(o.getExitDate())
+                              //                              && o.getExitDate().getYear() <= year)
+                              //                          || o.getExitDate().getYear() == year)
+                              //              .toList());
+                              // return asset1;
                             })
-                        .collect(Collectors.toList()));
+                        .toList());
               if (wallet.getType().equalsIgnoreCase("Trading")) {
                 Predicate<Operations> isNotTradingAndNotClosed =
                     operations1 ->
@@ -308,11 +421,15 @@ public class CryptoService {
                 operations.sort(c);
 
                 tradingLastBalance.updateAndGet(
-                    v -> ObjectUtils.isEmpty(operations) ? v : v + operations.get(0).getTrend());
+                    v ->
+                        ObjectUtils.isEmpty(operations) ? v : v + operations.getFirst().getTrend());
               }
               Predicate<Asset> hasEmptyStats =
                   asset -> asset.getHistory() == null || asset.getHistory().isEmpty();
-              List<Asset> filterAsset = wallet1.getAssets();
+              List<Asset> filterAsset =
+                  !ObjectToolkit.isNullOrEmpty(wallet1.getAssets())
+                      ? new ArrayList<>(wallet1.getAssets())
+                      : null;
               if (filterAsset != null && index.get() > 0) {
                 filterAsset.removeIf(hasEmptyStats);
                 wallet1.setAssets(filterAsset);
@@ -320,7 +437,7 @@ public class CryptoService {
               indexWallet.incrementAndGet();
               return wallet1;
             })
-        .collect(Collectors.toList());
+        .toList();
   }
 
   private void checkAndMapWalletInThePast(
@@ -334,57 +451,64 @@ public class CryptoService {
       // dell'ultima
       // data della lista dell'anno, il wallet non era ancora
       // cancellato
-      if (listFilter
-          .get(listFilter.size() - 1)
-          .getDate()
-          .isEqual(filterDateByYear.get(filterDateByYear.size() - 1))) {
+      if (listFilter.getLast().getDate().isEqual(filterDateByYear.getLast())) {
         wallet1.setDeletedDate(null);
       }
-      try {
-        cryptoMapper.mapWalletInThePast(wallet1);
-      } catch (UtilsException e) {
-        throw new RuntimeException(e);
-      }
+      CryptoMapper.mapWalletInThePast(wallet1);
     }
   }
 
-  private List<Asset> getCryptoAssetByYear(Boolean isResume, Integer year) {
-    return getCryptoAssetsList(isResume, year);
-  }
+  private List<Asset> getCryptoAssetsList(
+      Boolean isResume,
+      Integer year,
+      List<Wallet> walletList,
+      List<MarketData> marketData,
+      List<LocalDate> getAllDates) {
 
-  private List<Asset> getCryptoAsset(Boolean isResume) {
-    return getCryptoAssetsList(isResume, null);
-  }
-
-  private List<Asset> getCryptoAssetsList(Boolean isResume, Integer year) {
+    // Raccolgo tutti gli asset dai wallet
     List<Asset> assets =
-        objectMapper.convertValue(
-            Objects.requireNonNull(assetService.getAssets().getBody()).getData(),
-            new TypeReference<List<Asset>>() {});
+        walletList.stream()
+            .filter(wallet -> !ObjectToolkit.isNullOrEmpty(wallet.getAssets()))
+            .flatMap(wallet -> wallet.getAssets().stream())
+            .toList();
 
-    List<Asset> filterAsset = new ArrayList<>();
-    assets.forEach(
-        asset -> {
-          Asset newAsset = new Asset();
-          BeanUtils.copyProperties(asset, newAsset);
-          if (asset.getHistory() != null && !asset.getHistory().isEmpty() && !isResume) {
-            Stats stats = asset.getHistory().get(asset.getHistory().size() - 1);
-            newAsset.setHistory(List.of(stats));
-          }
-          // Se siamo in resume devo rimuovere gli Stats che non hanno history per l'anno passato
-          // come parametro
-          if (isResume && year != null && newAsset.getHistory() != null) {
-            Predicate<Stats> hasNotYearStats = stats -> stats.getDate().getYear() != year;
-            newAsset.getHistory().removeIf(hasNotYearStats);
-          }
-          filterAsset.add(newAsset);
-        });
+    // Mappa gli asset con marketData e getAllDates
+    assets = AssetMapper.mapAssetList(assets, marketData, getAllDates);
+
+    // Applica i filtri sugli asset
+    List<Asset> filteredAssets =
+        new ArrayList<>(
+            assets.stream()
+                .map(
+                    asset -> {
+                      Asset newAsset = new Asset();
+                      BeanUtils.copyProperties(asset, newAsset);
+
+                      if (asset.getHistory() != null && !asset.getHistory().isEmpty()) {
+                        if (!isResume) {
+                          // Se non è resume, mantieni solo l'ultima entry della history
+                          Stats latestStats = asset.getHistory().getLast();
+                          newAsset.setHistory(List.of(latestStats));
+                        } else if (year != null) {
+                          // Se è resume e un anno è specificato, filtra gli Stats per l'anno dato
+                          List<Stats> filteredHistory =
+                              asset.getHistory().stream()
+                                  .filter(stats -> stats.getDate().getYear() == year)
+                                  .toList();
+                          newAsset.setHistory(filteredHistory);
+                        }
+                      }
+
+                      return newAsset;
+                    })
+                .toList());
+
+    // Se è resume, rimuovi gli asset senza history
     if (isResume) {
-      Predicate<Asset> hasEmptyStats =
-          asset -> asset.getHistory() == null || asset.getHistory().isEmpty();
-      filterAsset.removeIf(hasEmptyStats);
+      filteredAssets.removeIf(asset -> ObjectToolkit.isNullOrEmpty(asset.getHistory()));
     }
-    return filterAsset;
+
+    return filteredAssets;
   }
 
   private Double getAssetValue(List<MarketData> marketData, String symbol) {
