@@ -1,6 +1,6 @@
 package com.giova.service.moneystats.app.attachments;
 
-import com.giova.service.moneystats.exception.ExceptionMap;
+import com.giova.service.moneystats.exception.config.ExceptionMap;
 import io.github.giovannilamarmora.utils.exception.ExceptionCode;
 import io.github.giovannilamarmora.utils.exception.UtilsException;
 
@@ -14,5 +14,13 @@ public class ImageException extends UtilsException {
 
   public ImageException(String message, String exceptionMessage) {
     super(DEFAULT_CODE, message, exceptionMessage);
+  }
+
+  public ImageException(ExceptionCode exceptionCode) {
+    super(exceptionCode, exceptionCode.getMessage());
+  }
+
+  public ImageException(ExceptionCode exceptionCode, String message) {
+    super(exceptionCode, message);
   }
 }
