@@ -1,14 +1,12 @@
 package com.giova.service.moneystats.app.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.giova.service.moneystats.app.category.dto.Category;
 import com.giova.service.moneystats.app.wallet.dto.Wallet;
+import java.time.LocalDate;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-import java.util.List;
 
 @Data
 @NoArgsConstructor
@@ -26,11 +24,13 @@ public class Dashboard {
   private Double lastStatsBalanceDifference;
 
   @JsonInclude(JsonInclude.Include.NON_EMPTY)
+  private List<Integer> yearsWalletStats;
+
+  @JsonInclude(JsonInclude.Include.NON_EMPTY)
   private List<LocalDate> statsWalletDays;
 
   @JsonInclude(JsonInclude.Include.NON_EMPTY)
-  private List<Category> categories;
-
-  @JsonInclude(JsonInclude.Include.NON_EMPTY)
   private List<Wallet> wallets;
+
+  private Boolean hasMoreRecords;
 }
