@@ -1,12 +1,12 @@
 package com.giova.service.moneystats.app.stats.entity;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.giova.service.moneystats.app.wallet.entity.WalletEntity;
-import com.giova.service.moneystats.authentication.entity.UserEntity;
 import com.giova.service.moneystats.crypto.asset.entity.AssetEntity;
 import io.github.giovannilamarmora.utils.generic.GenericEntity;
+import jakarta.persistence.*;
 import java.time.LocalDate;
-import javax.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -38,15 +38,16 @@ public class StatsEntity extends GenericEntity {
   @Column(name = "TREND", nullable = false)
   private Double trend;
 
-  @ManyToOne
-  @JoinColumn(name = "USER_ID", nullable = false)
-  private UserEntity user;
+  @Column(name = "USER_IDENTIFIER", nullable = false)
+  private String userIdentifier;
 
   @ManyToOne
   @JoinColumn(name = "ASSET_ID")
+  @JsonIgnore // 🔹 Evita la serializzazione ciclica
   private AssetEntity asset;
 
   @ManyToOne
-  @JoinColumn(name = "WALLET_ID", nullable = true)
+  @JoinColumn(name = "WALLET_ID")
+  @JsonIgnore // 🔹 Evita la serializzazione ciclica
   private WalletEntity wallet;
 }
