@@ -1,6 +1,6 @@
 package com.giova.service.moneystats.config;
 
-import com.giova.service.moneystats.authentication.entity.UserEntity;
+import com.giova.service.moneystats.authentication.dto.UserData;
 import io.github.giovannilamarmora.utils.config.OpenAPIConfig;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
@@ -8,10 +8,9 @@ import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
 import io.swagger.v3.oas.annotations.info.Info;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import io.swagger.v3.oas.annotations.servers.Server;
 import io.swagger.v3.oas.models.Paths;
 import java.util.Map;
-import org.springdoc.core.customizers.OpenApiCustomiser;
+import org.springdoc.core.customizers.OpenApiCustomizer;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -34,12 +33,12 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 public class AppConfig {
 
   @Bean
-  public UserEntity user() {
-    return new UserEntity();
+  public UserData userInfo() {
+    return new UserData();
   }
 
   @Bean
-  public OpenApiCustomiser applyStandardOpenAPIModifications() {
+  public OpenApiCustomizer applyStandardOpenAPIModifications() {
     return openApi -> {
       Paths paths = new Paths();
       openApi.getPaths().entrySet().stream()
