@@ -1,27 +1,26 @@
 package com.giova.service.moneystats.crypto.asset;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.giova.service.moneystats.app.stats.StatsService;
+import com.giova.service.moneystats.app.stats.StatsComponent;
 import com.giova.service.moneystats.app.wallet.WalletService;
 import com.giova.service.moneystats.app.wallet.dto.Wallet;
-import com.giova.service.moneystats.authentication.entity.UserEntity;
+import com.giova.service.moneystats.authentication.dto.UserData;
+import com.giova.service.moneystats.crypto.asset.database.AssetRepository;
 import com.giova.service.moneystats.crypto.asset.dto.Asset;
 import com.giova.service.moneystats.crypto.asset.entity.AssetEntity;
-import com.giova.service.moneystats.crypto.coinGecko.MarketDataService;
-import com.giova.service.moneystats.crypto.coinGecko.dto.MarketData;
-import io.github.giovannilamarmora.utils.exception.UtilsException;
+import com.giova.service.moneystats.crypto.marketData.MarketDataService;
+import com.giova.service.moneystats.crypto.marketData.dto.MarketData;
+import com.giova.service.moneystats.utilities.Utils;
+import io.github.giovannilamarmora.utils.context.TraceUtils;
 import io.github.giovannilamarmora.utils.generic.Response;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
 import io.github.giovannilamarmora.utils.interceptors.Logged;
-import io.github.giovannilamarmora.utils.interceptors.correlationID.CorrelationIdUtils;
+import io.github.giovannilamarmora.utils.utilities.ObjectToolkit;
+import jakarta.transaction.Transactional;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import javax.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,6 +28,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 
 @Service
 @Logged
@@ -36,112 +36,207 @@ import org.springframework.stereotype.Service;
 public class AssetService {
 
   private final Logger LOG = LoggerFactory.getLogger(this.getClass());
-  private final UserEntity user;
-  private final ObjectMapper mapper = new ObjectMapper().findAndRegisterModules();
+  private final UserData user;
   @Autowired private WalletService walletService;
-  @Autowired private IAssetDAO assetDAO;
-  @Autowired private AssetMapper assetMapper;
   @Autowired private MarketDataService marketDataService;
-  @Autowired private StatsService statsService;
+  @Autowired private StatsComponent statsComponent;
 
+  @Autowired private AssetRepository assetRepository;
+
+  /**
+   * Get Asset by his identifier
+   *
+   * @param identifier to be searched
+   * @return Asset
+   */
   @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
-  @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = Exception.class)
-  public ResponseEntity<Response> insertOrUpdateAssets(List<Wallet> wallets)
-      throws UtilsException, JsonProcessingException {
-    StringBuilder message = new StringBuilder("Assets ");
-    List<Wallet> walletRes = new ArrayList<>();
+  public ResponseEntity<Response> getAssetByIdentifier(String identifier) {
+    LOG.info("Getting asset {}", identifier);
+    List<AssetEntity> assetEntities =
+        assetRepository.findAllByIdentifierAndUserId(identifier, user.getIdentifier());
+    List<Asset> assets;
+    Asset asset = null;
 
-    wallets.forEach(
-        wallet -> {
-          try {
-            ResponseEntity<Response> saveWallet = walletService.insertOrUpdateWallet(wallet);
-            message
-                .append(wallet.getAssets().get(wallet.getAssets().size() - 1).getName())
-                .append(" ");
-            walletRes.add((Wallet) Objects.requireNonNull(saveWallet.getBody()).getData());
-          } catch (JsonProcessingException e) {
-            LOG.error("An Error happen during saving asset, message is {}", e.getMessage());
-            throw new AssetException(e.getMessage());
-          }
-        });
-    message.append("Successfully saved!");
+    String message;
 
-    Response response =
-        new Response(
-            HttpStatus.OK.value(),
-            message.toString(),
-            CorrelationIdUtils.getCorrelationId(),
-            Objects.requireNonNull(walletRes));
+    if (assetEntities.isEmpty()) {
+      message = "Asset " + identifier + " not found, insert new Asset to get it!";
+    } else {
+      List<LocalDate> getAllDates = statsComponent.getCryptoDistinctDates(user);
+
+      List<MarketData> marketData =
+          marketDataService.getMarketData(user.getSettings().getCryptoCurrency());
+      assets =
+          AssetMapper.mapAssetList(
+              AssetMapper.fromAssetEntitiesToAssets(assetEntities, marketData, getAllDates),
+              marketData,
+              getAllDates);
+      asset = assets.getFirst();
+      message = "Data for " + identifier;
+    }
+
+    Response response = new Response(HttpStatus.OK.value(), message, TraceUtils.getSpanID(), asset);
     return ResponseEntity.ok(response);
   }
 
+  /**
+   * Get All Asset
+   *
+   * @return Asset
+   */
   @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
-  @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = Exception.class)
-  public ResponseEntity<Response> insertOrUpdateAsset(Wallet wallet)
-      throws UtilsException, JsonProcessingException {
-
-    ResponseEntity<Response> saveWallet = walletService.insertOrUpdateWallet(wallet);
-
-    String message =
-        "Asset "
-            + wallet.getAssets().get(wallet.getAssets().size() - 1).getName()
-            + " Successfully saved!";
-
-    Response response =
-        new Response(
-            HttpStatus.OK.value(),
-            message,
-            CorrelationIdUtils.getCorrelationId(),
-            Objects.requireNonNull(saveWallet.getBody()).getData());
-    return ResponseEntity.ok(response);
-  }
-
-  @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
-  public ResponseEntity<Response> getAssets() {
-
-    List<AssetEntity> assetEntities = assetDAO.findAllByUserIdOrderByRank(user.getId());
+  public ResponseEntity<Response> getAssets(Boolean includeOperations) {
+    LOG.info("Getting all assets");
+    List<AssetEntity> assetEntities =
+        assetRepository.findAllByUserIdOrderByRank(user.getIdentifier());
     List<Asset> assets = new ArrayList<>();
-    List<LocalDate> getAllDates = statsService.getCryptoDistinctDates(user);
 
     String message = "";
     if (assetEntities.isEmpty()) {
       message = "Asset Empty, insert new Asset to get it!";
     } else {
+      if (!includeOperations)
+        assetEntities = assetEntities.stream().peek(asset -> asset.setOperations(null)).toList();
       List<MarketData> marketData =
           marketDataService.getMarketData(user.getSettings().getCryptoCurrency());
+
+      List<LocalDate> getAllDates = statsComponent.getCryptoDistinctDates(user);
       assets =
-          assetMapper.mapAssetList(
-              assetMapper.fromAssetEntitiesToAssets(assetEntities, marketData, getAllDates),
+          AssetMapper.mapAssetList(
+              AssetMapper.fromAssetEntitiesToAssets(assetEntities, marketData, getAllDates),
               marketData,
               getAllDates);
       message = "Found " + assets.size() + " Assets";
     }
 
     Response response =
-        new Response(HttpStatus.OK.value(), message, CorrelationIdUtils.getCorrelationId(), assets);
+        new Response(HttpStatus.OK.value(), message, TraceUtils.getSpanID(), assets);
     return ResponseEntity.ok(response);
   }
 
+  /**
+   * Add a new Crypto Asset
+   *
+   * @param wallet to be updated
+   * @return Asset
+   */
   @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
-  public ResponseEntity<Response> getAsset(String identifier) {
+  @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = Exception.class)
+  public ResponseEntity<Response> addAsset(Wallet wallet) {
+    LOG.info(
+        "Add asset {} for wallet {}", wallet.getAssets().getLast().getName(), wallet.getName());
 
-    List<Asset> assets =
-        mapper.convertValue(getAssets().getBody().getData(), new TypeReference<List<Asset>>() {});
-    Asset asset =
-        assets.stream()
-            .filter(a -> a.getIdentifier().equalsIgnoreCase(identifier))
-            .findFirst()
-            .orElseThrow(
-                () ->
-                    new AssetException(
-                        "An error happen during filtering assets, asset "
-                            + identifier
-                            + " was not found"));
+    Boolean isLiveWallet = Utils.isLiveWallet(null, user);
 
-    String message = "Data for " + identifier;
+    String message = "Asset " + wallet.getAssets().getLast().getName() + " Successfully saved!";
+
+    ResponseEntity<Response> saveWallet =
+        walletService.addOrUpdateWallet(wallet, isLiveWallet, message);
+
+    if (!ObjectUtils.isEmpty(saveWallet.getBody())) {
+      assetRepository.clearAllAssetsCache();
+    }
 
     Response response =
-        new Response(HttpStatus.OK.value(), message, CorrelationIdUtils.getCorrelationId(), asset);
+        new Response(
+            HttpStatus.OK.value(), message, TraceUtils.getSpanID(), saveWallet.getBody().getData());
+    return ResponseEntity.ok(response);
+  }
+
+  /**
+   * Update a Crypto Asset
+   *
+   * @param wallet to be updated
+   * @return Asset
+   */
+  @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
+  @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = Exception.class)
+  public ResponseEntity<Response> updateAsset(Wallet wallet) {
+    LOG.info(
+        "Update asset {} for wallet {}", wallet.getAssets().getLast().getName(), wallet.getName());
+
+    Boolean isLiveWallet = Utils.isLiveWallet(null, user);
+
+    String message = "Asset " + wallet.getAssets().getLast().getName() + " Successfully updated!";
+
+    ResponseEntity<Response> saveWallet =
+        walletService.addOrUpdateWallet(wallet, isLiveWallet, message);
+
+    if (!ObjectUtils.isEmpty(saveWallet.getBody())) {
+      assetRepository.clearAllAssetsCache();
+    }
+
+    Response response =
+        new Response(
+            HttpStatus.OK.value(), message, TraceUtils.getSpanID(), saveWallet.getBody().getData());
+    return ResponseEntity.ok(response);
+  }
+
+  /**
+   * Add a new Crypto Asset
+   *
+   * @param wallets to be updated
+   * @return Asset
+   */
+  @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
+  @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = Exception.class)
+  public ResponseEntity<Response> addAssets(List<Wallet> wallets) {
+    LOG.info("Add assets list started");
+    StringBuilder message = new StringBuilder("Assets ");
+    List<Wallet> walletRes = new ArrayList<>();
+
+    Boolean isLiveWallet = Utils.isLiveWallet(null, user);
+
+    wallets.forEach(
+        wallet -> {
+          ResponseEntity<Response> saveWallet =
+              walletService.addOrUpdateWallet(wallet, isLiveWallet, null);
+          message.append(wallet.getAssets().getLast().getName()).append(" ");
+          if (!ObjectToolkit.isNullOrEmpty(saveWallet.getBody()))
+            walletRes.add((Wallet) Objects.requireNonNull(saveWallet.getBody()).getData());
+        });
+    message.append("Successfully saved!");
+
+    if (!ObjectToolkit.isNullOrEmpty(walletRes)) {
+      assetRepository.clearAllAssetsCache();
+    }
+
+    Response response =
+        new Response(HttpStatus.OK.value(), message.toString(), TraceUtils.getSpanID(), walletRes);
+    return ResponseEntity.ok(response);
+  }
+
+  /**
+   * Update a Crypto Assets
+   *
+   * @param wallets to be updated
+   * @return Asset
+   */
+  @LogInterceptor(type = LogTimeTracker.ActionType.SERVICE)
+  @Transactional(value = Transactional.TxType.REQUIRED, rollbackOn = Exception.class)
+  public ResponseEntity<Response> updateAssets(List<Wallet> wallets) {
+    LOG.info("Update assets list started");
+    StringBuilder message = new StringBuilder("Assets ");
+    List<Wallet> walletRes = new ArrayList<>();
+
+    Boolean isLiveWallet = Utils.isLiveWallet(null, user);
+
+    wallets.forEach(
+        wallet -> {
+          ResponseEntity<Response> saveWallet =
+              walletService.addOrUpdateWallet(wallet, isLiveWallet, null);
+          message.append(wallet.getAssets().getLast().getName()).append(" ");
+          if (!ObjectToolkit.isNullOrEmpty(saveWallet.getBody()))
+            walletRes.add((Wallet) Objects.requireNonNull(saveWallet.getBody()).getData());
+        });
+    message.append("Successfully updated!");
+
+    if (!ObjectToolkit.isNullOrEmpty(walletRes)) {
+      assetRepository.clearAllAssetsCache();
+    }
+
+    Response response =
+        new Response(HttpStatus.OK.value(), message.toString(), TraceUtils.getSpanID(), walletRes);
     return ResponseEntity.ok(response);
   }
 }
