@@ -1,11 +1,12 @@
 package com.giova.service.moneystats.crypto.operations;
 
-import com.giova.service.moneystats.authentication.entity.UserEntity;
+import com.giova.service.moneystats.authentication.dto.UserData;
 import com.giova.service.moneystats.crypto.asset.entity.AssetEntity;
 import com.giova.service.moneystats.crypto.operations.dto.Operations;
 import com.giova.service.moneystats.crypto.operations.entity.OperationsEntity;
 import io.github.giovannilamarmora.utils.interceptors.LogInterceptor;
 import io.github.giovannilamarmora.utils.interceptors.LogTimeTracker;
+import io.github.giovannilamarmora.utils.utilities.ObjectToolkit;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.beans.BeanUtils;
@@ -15,24 +16,9 @@ import org.springframework.stereotype.Component;
 public class OperationsMapper {
 
   @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
-  public Operations fromOperationEntityToDTO(OperationsEntity operationsEntity) {
-    Operations operations = new Operations();
-    BeanUtils.copyProperties(operationsEntity, operations);
-    return operations;
-  }
-
-  @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
-  public OperationsEntity fromOperationDTOToEntity(
-      Operations operations, UserEntity user, AssetEntity asset) {
-    OperationsEntity operationsEntity = new OperationsEntity();
-    BeanUtils.copyProperties(operations, operationsEntity);
-    operationsEntity.setUser(user);
-    operationsEntity.setAsset(asset);
-    return operationsEntity;
-  }
-
-  @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
-  public List<Operations> fromOperationsEntitiesToDTOS(List<OperationsEntity> operationsEntities) {
+  public static List<Operations> fromOperationsEntitiesToDTOS(
+      List<OperationsEntity> operationsEntities) {
+    if (ObjectToolkit.isNullOrEmpty(operationsEntities)) return null;
     return operationsEntities.stream()
         .map(
             operationsEntity -> {
@@ -44,17 +30,35 @@ public class OperationsMapper {
   }
 
   @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
-  public List<OperationsEntity> fromOperationDTOSToEntities(
-      List<Operations> operations, UserEntity user, AssetEntity asset) {
+  public static List<OperationsEntity> fromOperationDTOSToEntities(
+      List<Operations> operations, UserData user, AssetEntity asset) {
+    if (ObjectToolkit.isNullOrEmpty(operations)) return null;
     return operations.stream()
         .map(
             operation -> {
               OperationsEntity operationsEntity = new OperationsEntity();
               BeanUtils.copyProperties(operation, operationsEntity);
-              operationsEntity.setUser(user);
+              operationsEntity.setUserIdentifier(user.getIdentifier());
               operationsEntity.setAsset(asset);
               return operationsEntity;
             })
         .collect(Collectors.toList());
+  }
+
+  @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
+  public Operations fromOperationEntityToDTO(OperationsEntity operationsEntity) {
+    Operations operations = new Operations();
+    BeanUtils.copyProperties(operationsEntity, operations);
+    return operations;
+  }
+
+  @LogInterceptor(type = LogTimeTracker.ActionType.MAPPER)
+  public OperationsEntity fromOperationDTOToEntity(
+      Operations operations, UserData user, AssetEntity asset) {
+    OperationsEntity operationsEntity = new OperationsEntity();
+    BeanUtils.copyProperties(operations, operationsEntity);
+    operationsEntity.setUserIdentifier(user.getIdentifier());
+    operationsEntity.setAsset(asset);
+    return operationsEntity;
   }
 }
